requirements.txt
black==22.12.0
click==8.1.3
colorama==0.4.6
dataclass-wizard==0.22.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
pycodestyle==2.10.0
Pymem==1.10.0
XInput-Python==0.4.0
